+++ packages/react-magma-icons/CHANGELOG.md
@@ -3,21 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.0-alpha.3 (2021-02-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+## 2.0.0-alpha.2 (2021-02-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+## 2.0.0-alpha.1 (2021-02-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.11-alpha.0 (2021-02-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
 ### 1.0.10 (2021-02-18)
 
 **Note:** Version bump only for package react-magma-icons
 
-
-
-
-
 ### 1.0.10-alpha.0 (2021-02-18)
 
 **Note:** Version bump only for package react-magma-icons
-
-
-
-
 
 ### 1.0.9 (2021-02-18)